+++ src/Comodo.jl
@@ -11,9 +11,19 @@ import Interpolations # E.g. for resampling curves
 import BSplineKit
 import QuadGK
 import Distances
-import DelaunayTriangulation
+import DelaunayTriangulation # For regiontrimesh
+import StaticArrays # For volumetric mesh definitions
 
 include("functions.jl")
+
+# Export imported packages
+export GeometryBasics
+
+# Export types
+export Nhedron, tet4, tet10, hex8, hex20, penta6 # Volumetric elements (polyhedra)
+
+# Export types/structs
+export ConnectivitySet
 
 # Export functions
 export comododir, elements2indices, hexbox, mindist, dist, lerp
@@ -29,20 +39,15 @@ export con_edge_face, con_edge_edge, con_face_edge, con_face_face, con_face_face
 export con_vertex_face, con_vertex_vertex, con_vertex_simplex, meshconnectivity
 export simplexcenter, vertex2simplexdata, simplex2vertexdata
 export circlepoints, loftlinear, trisurfslice
-export wrapindex, edgeangles, count_edge_face, boundaryedges, edges2curve
+export edgeangles, count_edge_face, boundaryedges, edges2curve
 export pointspacingmean, extrudecurve, meshgroup, ray_triangle_intersect
 export distmarch, mesh_curvature_polynomial, curve_length, evenly_sample, evenly_space
 export invert_faces, kabsch_rot, sweeploft, loftpoints2surf, revolvecurve
 export batman, tridisc, triplate, regiontrimesh, scalesimplex, subcurve, dualclad
+export tet2hex, element2faces, subhex, rhombicdodecahedron
 
 # Export functions: Visualization related
 export slidercontrol, slider2anim, dirplot, normalplot
-
-# Export types/structs
-export ConnectivitySet
-
-# Export imported packages
-export GeometryBasics
 
 end # module