+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Implement polyhedron types (`hex8`, `tet4`, etc)
+- Removed `wrapindex` function in favour of `mod1`
+- Added handling of boundary edges to `subquad` and `subtri`
 - Moved excess data from assets folder to dedicated repository